Refactoring server application service to enable modularity
- Use callbacks
- Provide app, db, config variables outside of server.js
- Pass Mongoose NodeJS ES6 promises instead of internal, deprecated
mpromise

From most parts work from
https://github.com/meanjs/mean/commit/edb62344bc4af3cb28c3469dc7463c13a2
fc76a1
